Move thumbnail-state to shared module
