+++ libs/roblox/src/index.ts
@@ -1,2 +1,3 @@
 // Export enums
-export { default as AssetType } from './enums/assetType';
+export { default as AssetType } from './enums/asset-type';
+export { default as ThumbnailState } from './enums/thumbnail-state';
