+++ services/web-app/src/js/hooks/useAuthenticatedUser.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import ThumbnailState from '../enums/thumbnailState';
+import { ThumbnailState } from 'roblox';
 import AuthenticatedUser from '../types/authenticated-user';
 import Thumbnail from '../types/thumbnail';
 import User from '../types/user';
