+++ src/js/services/thumbnails/batchProcessor.ts
@@ -1,7 +1,7 @@
 import { Batch, BatchItem } from '@tix-factory/batch';
 import Thumbnail from '../../types/thumbnail';
 import ThumbnailType from '../../enums/thumbnailType';
-import ThumbnailState from '../../enums/thumbnailState';
+import { ThumbnailState } from 'roblox';
 
 type ThumbnailBatchItem = {
   type: ThumbnailType;
