+++ src/js/services/thumbnails/index.ts
@@ -1,4 +1,4 @@
-import ThumbnailState from '../../enums/thumbnailState';
+import { ThumbnailState } from 'roblox';
 import ThumbnailType from '../../enums/thumbnailType';
 import Thumbnail from '../../types/thumbnail';
 import ExpirableDictionary from '../../utils/expireableDictionary';
